fix: restore multiple wiki support
when i removed the duplicate widget.widgetname.enabled keys, this broke
the ability of using the multiple documents feature. for now, the fix is
to make an exception for wikis and keep the other widget api's rigid, as
they write to disk through data/ and that colision (say, multiple
network widgets) would be bad.

longer term, this will have to be extended in two ways:

- widgets needing to be defensive of other activity on data, or
  intelligent about it through race-blocking locks, should gracefully
  handle it. users may want to stack charts through multiple metrics
  widgets: only one write is still needed.

- federation, i.e. spinning up monitor@metrics client for multiple
  monitor@machines metrics widgets will require this encapsulation
  anyway (TODO)