wip: add demo mode
adds a demo mode that allows monitorat to be served in a read-only state
and updates each widget's api to cutoff endpoints that would otherwise
spam the server.

disables:
- lazy reload from config change
- any cgi that calls systemd, docker
- actual disc reads and live system metrics

this is made possible by introducing a general snapshot.jsonl which
stores the monitor@ measured vars produced on refresh. this
functionality is needed for the demo because we don't want user hugs
spamming systemd, docker compose, etc on an actual live server shared
with the public.

this data storage coincidentally provides useful metadata and telemetry
about the system during interaction time. new widgets are encouraged to
join one-shot data to this log through their api.py, in addition to
whatever interval logging their code may do.